// src/shared/api/generated/__swagger_client/models/ReceiptLineRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ReceiptLineRequest = {
    itemId: string;
    unitId: string;
    warehouseZoneId?: string;
    warehouseCellId?: string;
    quantity: number;
    price?: number;
    amount?: number;
};